requirements.txt
absl-py==2.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.2.0
grpcio==1.64.1
gspread==6.1.2
h5py==3.11.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.82
packaging==24.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.5.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
triton==2.3.1
typing_extensions==4.9.0
urllib3==2.2.1
wandb==0.17.1
Werkzeug==3.0.3
